EventCommon was moved to SetupConfiguration.h to avoid dependencies on Go4 headers
