+++ setupconfigcppwrapper/SetupConfigCppWrapperLinkDef.h
@@ -5,6 +5,7 @@
 #pragma link off all functions;
 
 #pragma link C++ class SetupConfiguration+;
+#pragma link C++ class EventCommon+;
 #pragma link C++ class ElectronicsChars+;
 
 #endif // __CINT__
